refactor: rename Authenticate to TriggerProcedure
